+++ src/MapIterator.php
@@ -14,6 +14,8 @@ declare(strict_types=1);
 namespace League\Csv;
 
 use ArrayIterator;
+use Iterator;
+use IteratorAggregate;
 use IteratorIterator;
 use Traversable;
 
@@ -44,5 +46,22 @@ final class MapIterator extends IteratorIterator
     public function current(): mixed
     {
         return ($this->callable)(parent::current(), parent::key());
+    }
+
+    public static function toIterator(iterable $value): Iterator
+    {
+        if ($value instanceof IteratorAggregate) {
+            $value = $value->getIterator();
+        }
+
+        return match (true) {
+            $value instanceof Iterator => $value,
+            $value instanceof Traversable => (function () use ($value): Iterator {
+                foreach ($value as $offset => $record) {
+                    yield $offset => $record;
+                }
+            })(),
+            default => new ArrayIterator($value),
+        };
     }
 }