+++ src/Query/Constraint/Offset.php
@@ -13,13 +13,11 @@ declare(strict_types=1);
 
 namespace League\Csv\Query\Constraint;
 
-use ArrayIterator;
 use CallbackFilterIterator;
 use Closure;
 use Iterator;
-use IteratorIterator;
+use League\Csv\MapIterator;
 use League\Csv\Query;
-use Traversable;
 
 /**
  * Enable filtering a record based on its offset.
@@ -72,10 +70,6 @@ final class Offset implements Query\Predicate
 
     public function filter(iterable $value): Iterator
     {
-        return new CallbackFilterIterator(match (true) {
-            $value instanceof Iterator => $value,
-            $value instanceof Traversable => new IteratorIterator($value),
-            default => new ArrayIterator($value),
-        }, $this);
+        return new CallbackFilterIterator(MapIterator::toIterator($value), $this);
     }
 }