+++ src/Query/Constraint/TwoColumns.php
@@ -13,16 +13,14 @@ declare(strict_types=1);
 
 namespace League\Csv\Query\Constraint;
 
-use ArrayIterator;
 use CallbackFilterIterator;
 use Closure;
 use Iterator;
-use IteratorIterator;
+use League\Csv\MapIterator;
 use League\Csv\Query\Predicate;
 use League\Csv\Query\QueryException;
 use League\Csv\Query\Row;
 use ReflectionException;
-use Traversable;
 
 use function array_filter;
 use function is_array;
@@ -94,10 +92,6 @@ final class TwoColumns implements Predicate
 
     public function filter(iterable $value): Iterator
     {
-        return new CallbackFilterIterator(match (true) {
-            $value instanceof Iterator => $value,
-            $value instanceof Traversable => new IteratorIterator($value),
-            default => new ArrayIterator($value),
-        }, $this);
+        return new CallbackFilterIterator(MapIterator::toIterator($value), $this);
     }
 }